comment image crud logic v1

+++ src/main/java/com/example/highteenday_backend/services/domain/CommentMediaService.java
@@ -30,31 +30,33 @@ public class CommentMediaService {
         s3Service.deleteUserTmp(userId);
     }
 
+    //tmp 없이 바로 comment/commentId/postId로 저장
     @Transactional
     public void processUpdateCommentMedia(Long userId, Comment comment, RequestCommentDto dto){
         if (dto.getUrl().isEmpty()){ //이미지 삭제
             String deleteUrl = comment.getS3Url();
-
-            String deleteKey = s3Service.getKeyByUrl(deleteUrl);
-            s3Service.delete(deleteKey);
-            mediaService.deleteMediaByUrl(deleteUrl);
-
-            comment.updateImage(null);
+            if(!deleteUrl.isEmpty()){
+                String deleteKey = s3Service.getKeyByUrl(deleteUrl);
+                s3Service.delete(deleteKey);
+                mediaService.deleteMediaByUrl(deleteUrl);
+                comment.updateImage(null);
+            }
 
         } else if(!dto.getUrl().equals(comment.getS3Url())) { //이미지 업데이트
+            if(!comment.getS3Url().isEmpty()){ //기존 이미지 있을때 기존 이미지 제거
+                String pastKey = s3Service.getKeyByUrl(comment.getS3Url());
+                s3Service.delete(pastKey);
+                mediaService.deleteMediaByUrl(comment.getS3Url());
+            }
+            //새로운 s3,media 저장
             String finalUrl = s3Service.copyToFinalLocation(dto.getUrl(), comment.getId(), MediaOwner.COMMENT);
-            //새로운 s3 저장,기존+tmp s3 삭제
-            s3Service.deleteUserTmp(userId);
-            String pastKey = s3Service.getKeyByUrl(comment.getS3Url());
-            s3Service.delete(pastKey);
-            //새로운 media 저장, 기존 media 삭제
-            mediaService.deleteMediaByUrl(comment.getS3Url());
             String newKey = s3Service.getKeyByUrl(finalUrl);
             FileInfo newFileInfo = s3Service.getFileInfo(newKey);
             Media media = mediaService.createMedia(newFileInfo);
             media.setComment(comment);
-            //comment의 url 변경
             comment.updateImage(finalUrl);
+
+
         }
 
     }
